// src/modules/healthIntegeration/enum/HealthConnect.enum.ts
export enum HealthDataType {
  ACTIVE_CALORIES = 'ACTIVE_CALORIES',
  BASAL_METABOLIC_RATE = 'BASAL_METABOLIC_RATE',
  BLOOD_PRESSURE = 'BLOOD_PRESSURE',
  CYCLING_CADENCE = 'CYCLING_CADENCE',
  DISTANCE = 'DISTANCE',
  ELEVATION_GAINED = 'ELEVATION_GAINED',
  EXERCISE_SESSION = 'EXERCISE_SESSION',
  FLOORS_CLIMBED = 'FLOORS_CLIMBED',
  HEART_RATE = 'HEART_RATE',
  HEIGHT = 'HEIGHT',
  HYDRATION = 'HYDRATION',
  MINDFULNESS_SESSION = 'MINDFULNESS_SESSION',
  NUTRITION = 'NUTRITION',
  POWER = 'POWER',
  RESTING_HEART_RATE = 'RESTING_HEART_RATE',
  SKIN_TEMPERATURE = 'SKIN_TEMPERATURE',
  SLEEP_SESSION = 'SLEEP_SESSION',
  SPEED = 'SPEED',
  STEPS = 'STEPS',
  STEPS_CADENCE = 'STEPS_CADENCE',
  TOTAL_CALORIES_BURNED = 'TOTAL_CALORIES_BURNED',
  WEIGHT = 'WEIGHT',
  WHEELCHAIR_PUSHES = 'WHEELCHAIR_PUSHES',
}